cloudsql-proxy: Change to DaemonSet & remove autoscaler
We were having trouble with "MySQL has gone away" errors due to this service scaling up/down too frequently.

One proxy per node seems to be proving sufficient for our needs, so it may as well live as a DaemonSet. I expect this would be CPU bound, so if we do end up getting bottlenecked by this we can scale it vertically instead.