memcache: Improve reliability
Memcache was being a PITA due to "Too many connection" errors. This commit over-solves this, by scaling in both directions and increasing the amount of connections memcache can handle.

This should guarantee reliability, at the expense of excessive resource reservation. But resource usage is something we can observe & tune with time.