mediawiki: Vertically scale to reduce compute bottlenecks
Running with low performance limits worked fantastically for casual browsing, with a hot parsercache.

But it performed terribly when MediaWiki had to generate pages from scratch, and on every page load - where MediaWiki was dynamically generating the sidebar.

This bottleneck wasn't apparent in testing, in a single-user scenario. But under production traffic, it's been wreaking havoc. MediaWiki would perform fine when the stack was restarted, but performance would slowly degrade as Apache's requests queue would fill. Kubernetes was supposed to alleviate this by horizontally scaling, which should have led to a massive over-provisioning of pods, however the moment individual pods hit Apache's queue the pods would no longer return performance metrics, so Kubernetes had no insight to determine whether or not it should scale.

It turns out MediaWiki needs quite a lot of compute power to generate pages, so we're massively switching around our scaling strategy here. We're going from lots of low-performing pods, to a few high-performing pods.

This handles the same amount of traffic, but gives MediaWiki the juice it needs to generate pages quickly, and thus not cause a slow fill of the Apache process queue.

Alongside this change, we've had to make a similar update to the machines we're using in Google Cloud, from many 8-core machines  (which wouldn't hold one of these new MediaWiki pods) to 32-core machines (which can hold 3; with room for more).