Fix caching problems
Far as my reading suggests, using multiple memcache servers with PHP is supposed to just work a-ok. The extension/library/whatever automatically shards, so long as every instance of that library knows all the servers.

MediaWiki wraps all this in its own code, however, which just connects to a _single_ memcache instance every time it needs to do cache operations. This means MediaWiki was talking to a separate cache server every time it was getting, saving, or purging an item – hence our thoroughly inconsistent cache behaviour.

mcrouter is a tool that fixes this, by acting as a dedicated proxy to memcache outside of the application. Far as MediaWiki is concerned, its talking to one memcache server. mcrouter automatically shards and distributes any operations between our 10 memcache servers