Let prod mediawiki scale down to 2 instances
We're spending a ton of money on compute and its sitting idle most of the time. I introduced a shit ton of slack in the infra after some reliability incidents, but with recent upgrades and improvements we're much stabler; think we can cut some of this slack away to save pennies.